requirements.txt
asyncio
aiohttp
aiohttp_route_middleware
aiohttp_cors
PyJWT
pytest-aiohttp
tensorflow
